+++ src/app/app.component.ts
@@ -1,8 +1,7 @@
-import { formatCurrency } from '@angular/common';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
-import { map } from 'rxjs';
+
 import { IProductCreate } from './model/products';
+import { ProductsService } from './service/products.service';
 
 @Component({
   selector: 'app-root',
@@ -14,7 +13,7 @@ export class AppComponent implements OnInit {
   allProducts: IProductCreate[] = [];
   isFetching: boolean = false;
 
-  constructor(private http: HttpClient) {}
+  constructor(private productService: ProductsService) {}
 
   ngOnInit() {
     this.fetchProducts();
@@ -25,57 +24,22 @@ export class AppComponent implements OnInit {
   }
 
   onProductCreate(products: IProductCreate) {
-    console.log(products);
-    const hHeaders = new HttpHeaders({ myHeader: 'graygoo' });
-    this.http
-      .post<{ name: string }>(
-        'https://ngfeatures-general-concepts-default-rtdb.firebaseio.com/product.json',
-        products,
-        { headers: hHeaders }
-      )
-      .subscribe((res) => console.log(res));
+    this.productService.createProduct(products);
   }
 
   private fetchProducts() {
     this.isFetching = true;
-    this.http
-      .get<{ [key: string]: IProductCreate }>( //IProductCreate
-        'https://ngfeatures-general-concepts-default-rtdb.firebaseio.com/product.json'
-      )
-      .pipe(
-        map((res) => {
-          const products = [];
-          for (const key in res) {
-            // console.log('key: ', key);
-            // console.log('res: ', res);
-            if (res.hasOwnProperty(key)) {
-              products.push({ ...res[key], id: key });
-            }
-          }
-          return products;
-        })
-      )
-      .subscribe((products) => {
-        this.isFetching = false;
-        console.log(products);
-        this.allProducts = products;
-      });
+    this.productService.fetchProduct().subscribe((products) => {
+      this.isFetching = false;
+      console.log(products);
+      this.allProducts = products;
+    });
   }
 
   onDeleteProduct(id: string) {
-    this.http
-      .delete(
-        'https://ngfeatures-general-concepts-default-rtdb.firebaseio.com/product/' +
-          id +
-          '.json'
-      )
-      .subscribe();
+    this.productService.deleteProduct(id);
   }
   onDeleteAllProducts() {
-    this.http
-      .delete(
-        'https://ngfeatures-general-concepts-default-rtdb.firebaseio.com/product.json'
-      )
-      .subscribe();
+    this.productService.deleteAllProducts();
   }
 }
